refactor: rename package to geno_tools, bundle runtime dirs inside it
Rename the Python package genotools/ -> geno_tools/ (underscore — valid
import name matching the geno-tools CLI/repo). Updated all 3 entry points
and packages.find in pyproject.toml, every import across the package and
test suite, and conftest monkeypatch targets. The genotools.yaml manifest
filename is intentionally preserved.

Move the runtime dirs into the package as one self-contained tree:
  hooks/ scripts/ config/  ->  geno_tools/{hooks,scripts,config}/
and fix every path that pointed at them:
- bootstrap.sh: plugin-root derivation (now two levels deep) + config path
- hooks.json: bootstrap command path
- .claude-plugin/plugin.json: declare "hooks": "./geno_tools/hooks/hooks.json"
  so the relocated SessionStart hook is still discovered
- tests/test_plugin.py, README, SKILL.md, docs.py, compile_skill_docs.py paths

Tests: 134 pass; the 5 failures are pre-existing on main (missing
init-geno-dir.sh + a hook-script assertion that flags non-file commands),
unchanged by this refactor.
