+++ geno_tools/docs.py
@@ -10,7 +10,7 @@ and generates MkDocs-compatible markdown pages at multiple zoom levels:
   Level 5 — Rationale: (placeholder, generated by LLM during docs compilation)
 
 For full ecosystem compilation (workspace + install dir discovery, individual
-skill pages, ecosystem overview), use scripts/compile_skill_docs.py instead.
+skill pages, ecosystem overview), use geno_tools/scripts/compile_skill_docs.py instead.
 This module discovers from ~/.geno-tools/ only.
 """
 
@@ -24,7 +24,7 @@ from textwrap import dedent, indent
 
 import yaml
 
-from genotools.paths import ROOT
+from geno_tools.paths import ROOT
 
 CATEGORIES: dict[str, list[str]] = {
     "Core": ["geno-tools", "geno-agents", "geno-notes", "geno-mon", "geno-msg"],
