+++ geno_tools/hooks/hooks.json
@@ -6,7 +6,7 @@
         "hooks": [
           {
             "type": "command",
-            "command": "${CLAUDE_PLUGIN_ROOT}/scripts/bootstrap.sh",
+            "command": "${CLAUDE_PLUGIN_ROOT}/geno_tools/scripts/bootstrap.sh",
             "timeout": 30000
           },
           {
