+++ pyproject.toml
@@ -13,12 +13,12 @@ dependencies = ["pyyaml>=6.0"]
 test = ["pytest>=8.0"]
 
 [project.scripts]
-geno-tools = "genotools.cli:main"
-geno-trace = "genotools.trace:main"
-geno-docs = "genotools.docs:main"
+geno-tools = "geno_tools.cli:main"
+geno-trace = "geno_tools.trace:main"
+geno-docs = "geno_tools.docs:main"
 
 [tool.setuptools.packages.find]
-include = ["genotools*"]
+include = ["geno_tools*"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
